backend/requirements.txt
# FastAPI และ dependencies หลัก
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database และ ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication และ Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration และ Environment
pydantic-settings==2.0.3

# HTTP Client (สำหรับ integration tests)
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
